feat: 클릭 이벤트를 addClickEventTo 함수를 재활용해서 주는 식으로 변경, 반복되는 엘리먼트 생성 함수 loop.js에 분리
