+++ util.js
@@ -1,5 +1,5 @@
-import { createPageNotYet, genreFunction, headerFunction } from "./createPage.js";
-import { dummy } from "./sources.js";
+import { genreFunction, headerFunction } from "./createPage.js";
+import { createContentForms } from "./loop.js";
 
 export function pickFromArr (arr) {
     const length = arr.length;
@@ -24,145 +24,6 @@ export function initPage(){
     contents.innerHTML = '';
 }
 
-// 반복되는 컴포넌트들을 생성하는 함수들
-export function createContentForms(num) {
-    let res = '';
-
-    for(let i = 0; i < num; i++){
-        res += `
-        <div class="content-wrap">
-            <div class="content-wrap-img-wrap">
-                <img src=${pickFromArr(dummy.contentFormImg)} class="content-wrap-img">
-                <div class="content-wrap-img-info">
-                    <div class="content-score">1위</div>
-                    <div class="content-score-icon">🕔</div>
-                </div>
-            </div>
-        
-            <div class="content-wrap-desc">
-                <div class="content-name">${pickFromArr(dummy.contentName)}</div>
-                <div class="content-status">
-                    <img src="./img/img6.png" class="img-view">
-                    <div class="desc">${randomView()}만명</div>
-                </div>
-            </div>
-        </div>`
-    }
-
-    return res;
-}
-
-export function createRankingContents(num){
-    let res = '';
-    let rank = 1;
-
-    for(let i = 0; i < num; i++){
-        res += `
-    <div class="ranking-content">
-        <div class="ranking-num">${rank}</div>
-        <div class="ranking-img-wrap">
-            <img src=${pickFromArr(dummy.rankingImg)} class="ranking-img">
-        </div>
-        <div class="ranking-desc">
-            <div class="ranking-desc-title">${pickFromArr(dummy.contentName)}</div>
-            <div class="ranking-desc-info">
-                <img src="./img/img6.png" class="img-view">
-                <div class="desc">${randomView()}만명</div>
-                <div class="division"></div>
-                <div class="desc">기다무웹툰</div>
-                <div class="division"></div>
-                <div class="desc">${pickFromArr(dummy.contentGenre)}</div>
-                <div class="division"></div>
-                <div class="desc">${pickFromArr(dummy.contentWriter)}</div>
-            </div>
-            <div class="ranking-desc-day">${pickFromArr(dummy.contentDays)} 연재</div>
-        </div>
-    </div>`
-
-    rank++
-    }
-
-    return res
-}
-
-export function createRankingContentsWithoutNum(num){
-    let res = '';
-
-    for(let i = 0; i < num; i++){
-        res += `
-    <div class="ranking-content">
-        <div class="ranking-img-wrap">
-            <img src=${pickFromArr(dummy.rankingImg)} class="ranking-img">
-        </div>
-        <div class="ranking-desc">
-            <div class="ranking-desc-title">${pickFromArr(dummy.contentName)}</div>
-            <div class="ranking-desc-info">
-                <img src="./img/img6.png" class="img-view">
-                <div class="desc">${randomView()}만명</div>
-                <div class="division"></div>
-                <div class="desc">기다무웹툰</div>
-                <div class="division"></div>
-                <div class="desc">${pickFromArr(dummy.contentGenre)}</div>
-                <div class="division"></div>
-                <div class="desc">${pickFromArr(dummy.contentWriter)}</div>
-            </div>
-            <div class="ranking-desc-day">${pickFromArr(dummy.contentDays)} 연재</div>
-        </div>
-    </div>`
-    }
-
-    return res
-}
-
-export function createHorizontalContents(num){
-    let res = '';
-
-    for(let i = 0; i < num; i++){
-        res += `
-    <div class="ranking-content">
-        <div class="horizontal-img-wrap">
-            <img src=${pickFromArr(dummy.horizontalImg)} class="horizontal-img">
-        </div>
-        <div class="ranking-desc">
-            <div class="ranking-desc-title">${pickFromArr(dummy.contentName)}</div>
-            <div class="desc" style="margin-top: 10px">${pickFromArr(dummy.contentDesc)}</div>
-            <div class="ranking-desc-info">
-                <img src="./img/img6.png" class="img-view">
-                <div class="desc">${randomView()}만명</div>
-                <div class="division"></div>
-                <div class="desc">${pickFromArr(dummy.contentWriter)}</div>
-            </div>
-        </div>
-    </div>`
-    }
-
-    return res
-}
-
-export function createContentFormsBigger(num){
-    let res = '';
-
-    for(let i = 0; i < num; i++){
-        res += `
-    <div class="content-wrap-bigger">
-        <div class="content-vertical-img-wrap">
-            <img src=${pickFromArr(dummy.contentFormBiggerImg)} class="content-vertical-img">
-        </div>
-
-        <div class="content-wrap-desc">
-            <div class="content-name">${pickFromArr(dummy.contentName)}</div>
-            <div class="content-status">
-                <img src="./img/img6.png" class="img-view">
-                <div class="desc">${randomView()}만명</div>
-            </div>
-        </div>
-    </div>`
-    }
-
-    return res
-}
-//
-
 export function addClickEventToElement(elementName, func){
     const element = document.querySelector(elementName);
 
@@ -175,16 +36,12 @@ export function selectedInit(target){
     })
 }
 
-export function addSelectedTo(element){
-    const elements = document.querySelector(element)
-
-    elements.addEventListener('click', e => {
-        selectedInit(e.target)
-        if(e.target.tagName === "LI"){
-            e.target.classList.add('selected')
-            resetFlexbox(e.target);
-        }
-    })
+export function addSelected (e) {
+    selectedInit(e.target)
+    if (e.target.tagName === "LI") {
+        e.target.classList.add('selected')
+        resetFlexbox(e.target);
+    }
 }
 
 export function combineFormats(arr){
@@ -210,25 +67,18 @@ export function resetFlexbox(target){
     }
 }
 
-export function addGenreNavEventTo (element) {
-    const el = document.querySelector(element);
+export function addHeadNavEvent (e) {
+    const header = e.target.dataset.header
 
-    el.addEventListener("click", e => {
-        const genre = e.target.dataset.genre;
-        if (genre) {
-            genreFunction[genre]()
-        }
-    })
+    if (header) {
+        headerFunction[header]();
+    }
 }
 
-export function addHeadNavEventTo (element) {
-    const el = document.querySelector(element);
+export function addGenreNavEvent (e) {
+    const genre = e.target.dataset.genre;
 
-    el.addEventListener("click", e => {
-        const header = e.target.dataset.header
-
-        if(header){
-            headerFunction[header]();
-        }
-    })
+    if (genre) {
+        genreFunction[genre]()
+    }
 }
